Add files via upload
In the course of the web-based system, we will design a website
to support one of the accounts on social networking sites
Instagram Where the account focuses on manual work,
especially drawing on pots with a special design requested by
the customer. Initially, when we design the site, there will be a
home page containing the official Logo And the welcome words
for the user, and then the user will choose the section from
which he wants to buy and put his products in the shopping cart
The goal of designing the site is to facilitate the ordering
process, search for products available to customers, and the
possibility of expressing their opinions while providing different
methods of payment upon completion of the order, and the
possibility of communicating with the business owner to request
assistance or inquiries. And to provide an easy-to-use user
interface, where the customer can remember what he wants to
do without the need to learn anything or take time for that.